third_party/blink/web_tests/virtual/stable/external/wpt/css/css-animations/parsing/animation-delay-shorthand-expected.txt

This is a testharness.js-based test.
Found 66 tests; 16 PASS, 50 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS e.style['animation-delay'] = "1s" should set the property value
PASS e.style['animation-delay'] = "-1s" should set the property value
FAIL e.style['animation-delay'] = "1s 2s" should set the property value assert_not_equals: property should be set got disallowed value ""
PASS e.style['animation-delay'] = "1s, 2s" should set the property value
FAIL e.style['animation-delay'] = "1s 2s, 3s" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-delay'] = "1s, 2s 3s" should set the property value assert_not_equals: property should be set got disallowed value ""
PASS e.style['animation-delay'] = "1s, 2s, 3s" should set the property value
FAIL e.style['animation-delay'] = "cover" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-delay'] = "contain" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-delay'] = "enter" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-delay'] = "exit" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-delay'] = "enter, exit" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-delay'] = "enter 0% enter 100%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-delay'] = "exit 0% exit 100%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-delay'] = "cover 0% cover 100%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-delay'] = "contain 0% contain 100%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-delay'] = "cover 50%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-delay'] = "contain 50%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-delay'] = "enter 50%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-delay'] = "exit 50%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-delay'] = "enter 50% 0s" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-delay'] = "0s enter 50%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-delay'] = "enter 50% exit 50%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-delay'] = "cover 50% enter 50%, contain 50% exit 50%" should set the property value assert_not_equals: property should be set got disallowed value ""
PASS e.style['animation-delay'] = "1s 2s 3s" should not set the property value
PASS e.style['animation-delay'] = "0s, 1s 2s 3s" should not set the property value
PASS e.style['animation-delay'] = "1s / 2s" should not set the property value
PASS e.style['animation-delay'] = "1s, 2px" should not set the property value
PASS e.style['animation-delay'] = "#ff0000" should not set the property value
PASS e.style['animation-delay'] = "red" should not set the property value
PASS e.style['animation-delay'] = "thing" should not set the property value
PASS e.style['animation-delay'] = "thing 0%" should not set the property value
PASS e.style['animation-delay'] = "thing 42%" should not set the property value
PASS e.style['animation-delay'] = "thing 100%" should not set the property value
PASS e.style['animation-delay'] = "thing 100px" should not set the property value
PASS e.style['animation-delay'] = "100% thing" should not set the property value
FAIL e.style['animation-delay'] = "1s 2s" should set animation-delay-end assert_equals: animation-delay-end should be canonical expected (string) "2s" but got (undefined) undefined
FAIL e.style['animation-delay'] = "1s 2s" should set animation-delay-start assert_equals: animation-delay-start should be canonical expected (string) "1s" but got (undefined) undefined
FAIL e.style['animation-delay'] = "1s 2s" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-delay'] = "1s" should set animation-delay-end assert_equals: animation-delay-end should be canonical expected (string) "0s" but got (undefined) undefined
FAIL e.style['animation-delay'] = "1s" should set animation-delay-start assert_equals: animation-delay-start should be canonical expected (string) "1s" but got (undefined) undefined
FAIL e.style['animation-delay'] = "1s" should not set unrelated longhands assert_equals: expected 0 but got 1
FAIL e.style['animation-delay'] = "cover" should set animation-delay-end assert_equals: animation-delay-end should be canonical expected (string) "cover 100%" but got (undefined) undefined
FAIL e.style['animation-delay'] = "cover" should set animation-delay-start assert_equals: animation-delay-start should be canonical expected (string) "cover 0%" but got (undefined) undefined
FAIL e.style['animation-delay'] = "cover" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-delay'] = "contain" should set animation-delay-end assert_equals: animation-delay-end should be canonical expected (string) "contain 100%" but got (undefined) undefined
FAIL e.style['animation-delay'] = "contain" should set animation-delay-start assert_equals: animation-delay-start should be canonical expected (string) "contain 0%" but got (undefined) undefined
FAIL e.style['animation-delay'] = "contain" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-delay'] = "enter 10% exit 20%" should set animation-delay-end assert_equals: animation-delay-end should be canonical expected (string) "exit 20%" but got (undefined) undefined
FAIL e.style['animation-delay'] = "enter 10% exit 20%" should set animation-delay-start assert_equals: animation-delay-start should be canonical expected (string) "enter 10%" but got (undefined) undefined
FAIL e.style['animation-delay'] = "enter 10% exit 20%" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-delay'] = "1s 2s, 3s 4s" should set animation-delay-end assert_equals: animation-delay-end should be canonical expected (string) "2s, 4s" but got (undefined) undefined
FAIL e.style['animation-delay'] = "1s 2s, 3s 4s" should set animation-delay-start assert_equals: animation-delay-start should be canonical expected (string) "1s, 3s" but got (undefined) undefined
FAIL e.style['animation-delay'] = "1s 2s, 3s 4s" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-delay'] = "1s 2s, 3s, 4s 5s" should set animation-delay-end assert_equals: animation-delay-end should be canonical expected (string) "2s, 0s, 5s" but got (undefined) undefined
FAIL e.style['animation-delay'] = "1s 2s, 3s, 4s 5s" should set animation-delay-start assert_equals: animation-delay-start should be canonical expected (string) "1s, 3s, 4s" but got (undefined) undefined
FAIL e.style['animation-delay'] = "1s 2s, 3s, 4s 5s" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-delay'] = "enter, exit" should set animation-delay-end assert_equals: animation-delay-end should be canonical expected (string) "enter 100%, exit 100%" but got (undefined) undefined
FAIL e.style['animation-delay'] = "enter, exit" should set animation-delay-start assert_equals: animation-delay-start should be canonical expected (string) "enter 0%, exit 0%" but got (undefined) undefined
FAIL e.style['animation-delay'] = "enter, exit" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-delay'] = "enter 0%, exit" should set animation-delay-end assert_equals: animation-delay-end should be canonical expected (string) "0s, exit 100%" but got (undefined) undefined
FAIL e.style['animation-delay'] = "enter 0%, exit" should set animation-delay-start assert_equals: animation-delay-start should be canonical expected (string) "enter 0%, exit 0%" but got (undefined) undefined
FAIL e.style['animation-delay'] = "enter 0%, exit" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-delay'] = "enter 0% 1s, 2s exit 50%" should set animation-delay-end assert_equals: animation-delay-end should be canonical expected (string) "1s, exit 50%" but got (undefined) undefined
FAIL e.style['animation-delay'] = "enter 0% 1s, 2s exit 50%" should set animation-delay-start assert_equals: animation-delay-start should be canonical expected (string) "enter 0%, 2s" but got (undefined) undefined
FAIL e.style['animation-delay'] = "enter 0% 1s, 2s exit 50%" should not set unrelated longhands assert_true: expected true got false
Harness: the test ran to completion.
